requirements.txt
tensorflow
opencv-python
numpy
streamlit
pillow
